mcp-server/requirements.txt
mcp
sqlalchemy
pymysql
python-dotenv
pytest
